chore: split out qs config types into their own file
* split out quickstart config types into their own file
* adjusted names of types to no longer use yaml, use config instead
* adjusted references to match new names & location of types
